moving FireAnalyzer function
