+++ src/WinTenDev.Zizi.Services/Telegram/TelegramService.cs
@@ -1899,45 +1899,6 @@ public class TelegramService
         await ScheduleKickJob(name);
     }
 
-    public async Task<StringAnalyzer> FireAnalyzer()
-    {
-        var settings = await GetChatSetting();
-        StringAnalyzer result = new();
-
-        if (!settings.EnableFireCheck)
-        {
-            Log.Information("Fire Check is disabled on ChatID '{ChatId}'", ChatId);
-            return result;
-        }
-
-        result = ChatService.FireAnalyzer(MessageOrEditedText);
-
-        if (!result.IsFired) return result;
-        var muteUntil = result.FireRatio * 1.33;
-        var untilDate = DateTime.Now.AddHours(muteUntil);
-
-        var sendText = result.ResultNote;
-
-        if (!await CheckUserPermission())
-        {
-            sendText += $"\nAnda di Mute sampai {untilDate} ";
-            await RestrictMemberAsync(FromId, until: untilDate);
-        }
-
-        var replyMarkup = new InlineKeyboardMarkup(
-            new[]
-            {
-                new[]
-                {
-                    InlineKeyboardButton.WithCallbackData("🧹 Hapus Debuff", $"un-restrict {FromId}")
-                }
-            }
-        );
-
-        await SendTextMessageAsync(sendText, replyMarkup: replyMarkup, scheduleDeleteAt: untilDate);
-        return result;
-    }
-
     public async Task ScheduleKickJob(StepHistoryName name)
     {
         Log.Information("Scheduling Job with name: {JobName}", name);
